refactor(workspace): separate gestures onboarding and navigation
