+++ src/features/workspace/styles/camera.css
@@ -1,0 +1,223 @@
+.camera-preview {
+  position: absolute;
+  top: 5.75rem;
+  right: var(--space-xl);
+  z-index: var(--z-floating-controls);
+  display: grid;
+  grid-template-columns: minmax(0, 1fr);
+  width: 10rem;
+  justify-items: end;
+  gap: var(--space-md);
+}
+
+.camera-preview__viewport {
+  position: relative;
+  display: grid;
+  width: 10rem;
+  aspect-ratio: 1;
+  place-items: center;
+  overflow: hidden;
+  color: var(--foreground);
+  background: var(--shell-sunken);
+  border: 2px solid var(--primary);
+  border-radius: 50%;
+}
+
+button.camera-preview__viewport {
+  padding: 0;
+  font: inherit;
+  appearance: none;
+  cursor: pointer;
+  transition:
+    border-color var(--duration-fast) var(--ease-out-quart),
+    box-shadow var(--duration-fast) var(--ease-out-quart),
+    transform var(--duration-fast) var(--ease-out-quart);
+}
+
+button.camera-preview__viewport:not(:disabled):hover {
+  border-color: var(--ring);
+  box-shadow: 0 0 0 0.25rem oklch(0.72 0.16 293 / 18%);
+}
+
+button.camera-preview__viewport:not(:disabled):active {
+  transform: scale(0.96);
+}
+
+button.camera-preview__viewport:focus-visible {
+  outline: 3px solid var(--ring);
+  outline-offset: 3px;
+}
+
+button.camera-preview__viewport:disabled {
+  cursor: default;
+}
+
+.camera-preview__video {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  transform: scaleX(-1);
+}
+
+.camera-preview__calibration-guide {
+  position: absolute;
+  inset: 12%;
+  pointer-events: none;
+  border: 2px dashed var(--primary);
+  border-radius: 50%;
+}
+
+.camera-preview__landmarks {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+}
+
+.camera-preview__loader {
+  width: 2rem;
+  height: 2rem;
+  color: var(--primary);
+  animation: camera-spin 800ms linear infinite;
+}
+
+.camera-preview__controls {
+  display: grid;
+  width: 100%;
+  justify-items: end;
+  gap: var(--space-sm);
+  color: var(--foreground);
+}
+
+.camera-preview__gesture-toast {
+  display: inline-flex;
+  width: max-content;
+  max-width: min(14rem, calc(100vw - 2 * var(--space-xl)));
+  min-height: 2.25rem;
+  align-items: center;
+  justify-self: end;
+  gap: var(--space-sm);
+  padding: var(--space-xs) var(--space-md);
+  color: var(--foreground);
+  font-size: 0.75rem;
+  font-weight: 600;
+  line-height: 1.5;
+  text-align: left;
+  white-space: normal;
+  background: var(--shell-raised);
+  border-radius: 999px;
+  box-shadow: 0 0.125rem 0.5rem oklch(0.06 0.01 286 / 22%);
+  transform-origin: top right;
+  animation: camera-gesture-toast 1650ms var(--ease-out-quart) both;
+}
+
+.camera-preview__gesture-toast[data-persistent] {
+  animation: camera-gesture-toast-enter 200ms var(--ease-out-quart) both;
+}
+
+.camera-preview__gesture-toast > svg {
+  width: 1rem;
+  height: 1rem;
+  flex: none;
+  color: var(--primary);
+}
+
+.camera-preview__gesture-toast > span {
+  min-width: 0;
+  text-wrap: pretty;
+}
+
+.camera-preview__gesture-toast[data-kind="uncertain"] > svg,
+.camera-preview__gesture-toast[data-kind="lost"] > svg {
+  color: var(--warning);
+}
+
+@keyframes camera-gesture-toast-enter {
+  from {
+    opacity: 0;
+    transform: translateY(-0.375rem) scale(0.96);
+  }
+}
+
+@keyframes camera-gesture-toast {
+  0% {
+    opacity: 0;
+    transform: translateY(-0.375rem) scale(0.96);
+  }
+
+  12%,
+  78% {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+
+  100% {
+    opacity: 0;
+    transform: translateY(-0.25rem) scale(0.98);
+  }
+}
+
+.camera-preview__message,
+.camera-preview__error {
+  display: flex;
+  width: 100%;
+  gap: var(--space-sm);
+}
+
+.camera-preview__message {
+  align-items: flex-start;
+  padding: var(--space-md);
+  color: var(--foreground);
+  font-size: 0.875rem;
+  line-height: 1.4;
+  background: var(--shell-raised);
+  border-radius: var(--radius-lg);
+  box-shadow: 0 0.25rem 0.5rem oklch(0.06 0.01 286 / 20%);
+}
+
+.camera-preview__message > svg {
+  flex: 0 0 auto;
+  width: 1rem;
+  height: 1rem;
+  margin-top: 0.125rem;
+  color: var(--primary);
+}
+
+.camera-preview__error {
+  flex-direction: column;
+  width: 18rem;
+  justify-self: end;
+}
+
+.camera-preview__device-field {
+  display: grid;
+  width: 18rem;
+  justify-self: end;
+  gap: var(--space-xs);
+  padding: var(--space-sm);
+  color: var(--foreground);
+  font-size: 0.75rem;
+  background: var(--shell-raised);
+  border-radius: var(--radius-lg);
+}
+
+.camera-preview__device-field select {
+  min-height: 2.75rem;
+  padding-inline: var(--space-sm);
+  color: var(--foreground);
+  background: var(--shell-sunken);
+  border: 1px solid var(--border);
+  border-radius: var(--radius-md);
+}
+
+.camera-preview__device-field select:focus-visible {
+  outline: 2px solid var(--ring);
+  outline-offset: 2px;
+}
+
+@keyframes camera-spin {
+  to {
+    transform: rotate(1turn);
+  }
+}
